\name{FAdist-package}
\Rdversion{1.1}
\alias{FAdist-package}
\alias{FAdist}
\docType{package}
\title{Distributions that are sometimes used in hydrology}
\description{This package contains several distributions that are sometimes useful in hydrology}

\author{
Francois Aucoin

Maintainer: Francois Aucoin <frank.aucoin@gmail.com>
}

\keyword{package}
